Crate gutenberg_post_parser

The Gutenberg post parser.

Gutenberg is a new post editor for the WordPress ecosystem. A post has always been HTML, and it continues to be. The difference is that the HTML is annotated. Like most annotation languages, it is located in comments, like this:

<h1>Famous post</h1>

<!-- wp:component {attributes: "as JSON"} -->
lorem ipsum
<!-- /wp:component -->

The parser analyses a post and generates an Abstract Syntax Tree (AST) of it.

Platforms and bindings

The parser aims at being used on different platforms, such as: Web within multiple browsers, Web applications like Electron, native applications like macOS, iOS, Windows, Linux etc.

Thus, the parser can be compiled as a binary, a static library, can be embedded in any Rust projects, can be compiled to WebAssembly, or can be compiled to NodeJS, and soon more.

This project uses Justfile as an alternative to Makefile. Every following command will use just, you might consider to install it. To learn about all the commands, just just --list.

Binary

To compile the parser to a binary, run:

$ just build-binary
$ echo -n '<!-- wp:foo {"bar": "qux"} /-->' | ./target/release/gutenberg-post-parser

Static library

To compile the parser to a static library, run:

$ just build-library
$ ls target/release/

WebAssembly

To compile the parser to a WebAssembly file, run:

$ just build-wasm
$ open bindings/wasm/index.html # for a demonstration

NodeJS

To compile the parser to a NodeJS native module, run:

$ just build-nodejs
$ node bindings/nodejs/lib/index.js # for a demonstration

Performance and guarantee

The parser guarantees to never copy the data in memory, which makes it fast and memory efficient.

License

The license is a classic BSD-3-Clause:

New BSD License

Copyright © Ivan Enderlin. All rights reserved.

Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:

Redistributions of source code must retain the above copyright notice, this list of conditions and the following disclaimer.

Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following disclaimer in the documentation and/or other materials provided with the distribution.

Neither the name of this project nor the names of its contributors may be used to endorse or promote products derived from this software without specific prior written permission.

THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDERS AND CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Modules

ast
The Abstract Syntax Tree (AST), i.e. the output of the parser.

combinators
Additional combinators specifically tailored for this parser.

parser
The Gutenberg post parser.

Macros

fold_into_vector_many0
fold_into_vector_many0!(I -> IResult<I,O>, R) => I -> IResult<I, R> is a wrapper around fold_many0! specifically designed for vectors.

take_till_terminated
take_till_terminated(S, C) is like take_till from but with a lookahead combinator C.

Functions

root
The root function represents the axiom of the grammar, i.e. the top rule.

Type Definitions

Input
Represent the type of a parser input.

InputElement
Represent the type of a parser input element. See Input for more information.